Expose CCR to the transport client
This commit exposes CCR to the transport client.
