+++ x-pack/plugin/ccr/src/test/java/org/elasticsearch/xpack/ccr/action/ShardFollowNodeTaskTests.java
@@ -13,6 +13,7 @@ import org.elasticsearch.index.shard.ShardNotFoundException;
 import org.elasticsearch.index.translog.Translog;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xpack.ccr.action.bulk.BulkShardOperationsResponse;
+import org.elasticsearch.xpack.core.ccr.ShardFollowNodeTaskStatus;
 
 import java.net.ConnectException;
 import java.nio.charset.StandardCharsets;
@@ -44,7 +45,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
     private List<List<Translog.Operation>> bulkShardOperationRequests;
     private BiConsumer<TimeValue, Runnable> scheduler = (delay, task) -> task.run();
 
-    private Consumer<ShardFollowNodeTask.Status> beforeSendShardChangesRequest = status -> {};
+    private Consumer<ShardFollowNodeTaskStatus> beforeSendShardChangesRequest = status -> {};
 
     private AtomicBoolean simulateResponse = new AtomicBoolean();
 
@@ -66,7 +67,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests, contains(new long[][]{
             {6L, 8L}, {14L, 8L}, {22L, 8L}, {30L, 8L}, {38L, 8L}, {46L, 8L}, {54L, 7L}}
         ));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(7));
         assertThat(status.lastRequestedSeqNo(), equalTo(60L));
     }
@@ -86,7 +87,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         task.innerHandleReadResponse(0L, 63L, generateShardChangesResponse(0, 63, 0L, 128L));
         assertThat(shardChangesRequests.size(), equalTo(0)); // no more reads, because write buffer is full
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(0));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
@@ -102,7 +103,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests.get(0)[0], equalTo(0L));
         assertThat(shardChangesRequests.get(0)[1], equalTo(8L));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.lastRequestedSeqNo(), equalTo(7L));
     }
@@ -140,7 +141,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests.size(), equalTo(0)); // no more reads, because task has been cancelled
         assertThat(bulkShardOperationRequests.size(), equalTo(0)); // no more writes, because task has been cancelled
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(0));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.lastRequestedSeqNo(), equalTo(15L));
@@ -164,7 +165,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests.size(), equalTo(0)); // no more reads, because task has been cancelled
         assertThat(bulkShardOperationRequests.size(), equalTo(0)); // no more writes, because task has been cancelled
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(0));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
@@ -211,7 +212,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         }
 
         assertFalse("task is not stopped", task.isStopped());
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.numberOfFailedFetches(), equalTo((long)max));
@@ -258,7 +259,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertTrue("task is stopped", task.isStopped());
         assertThat(fatalError, notNullValue());
         assertThat(fatalError.getMessage(), containsString("retrying failed ["));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.numberOfFailedFetches(), equalTo(11L));
@@ -299,7 +300,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
 
         assertTrue("task is stopped", task.isStopped());
         assertThat(fatalError, sameInstance(failure));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.numberOfFailedFetches(), equalTo(1L));
@@ -326,7 +327,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(bulkShardOperationRequests.size(), equalTo(1));
         assertThat(bulkShardOperationRequests.get(0), equalTo(Arrays.asList(response.getOperations())));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.mappingVersion(), equalTo(0L));
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
@@ -353,7 +354,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests.get(0)[0], equalTo(21L));
         assertThat(shardChangesRequests.get(0)[1], equalTo(43L));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
@@ -376,7 +377,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
 
         assertThat(shardChangesRequests.size(), equalTo(0));
         assertThat(bulkShardOperationRequests.size(), equalTo(0));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(0));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
@@ -399,7 +400,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests.get(0)[0], equalTo(0L));
         assertThat(shardChangesRequests.get(0)[1], equalTo(64L));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
@@ -441,7 +442,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(bulkShardOperationRequests.size(), equalTo(1));
         assertThat(bulkShardOperationRequests.get(0), equalTo(Arrays.asList(response.getOperations())));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.mappingVersion(), equalTo(1L));
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
@@ -466,7 +467,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(mappingUpdateFailures.size(), equalTo(0));
         assertThat(bulkShardOperationRequests.size(), equalTo(1));
         assertThat(task.isStopped(), equalTo(false));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.mappingVersion(), equalTo(1L));
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
@@ -492,7 +493,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(mappingVersions.size(), equalTo(1));
         assertThat(bulkShardOperationRequests.size(), equalTo(0));
         assertThat(task.isStopped(), equalTo(true));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.mappingVersion(), equalTo(0L));
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
@@ -511,7 +512,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
 
         assertThat(bulkShardOperationRequests.size(), equalTo(0));
         assertThat(task.isStopped(), equalTo(true));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.mappingVersion(), equalTo(0L));
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(0));
@@ -535,7 +536,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(bulkShardOperationRequests.size(), equalTo(1));
         assertThat(bulkShardOperationRequests.get(0), equalTo(Arrays.asList(response.getOperations())));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
@@ -553,7 +554,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(bulkShardOperationRequests.get(0), equalTo(Arrays.asList(response.getOperations()).subList(0, 64)));
         assertThat(bulkShardOperationRequests.get(1), equalTo(Arrays.asList(response.getOperations()).subList(64, 128)));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentWrites(), equalTo(2));
 
         task = createShardFollowTask(64, 1, 4, Integer.MAX_VALUE, Long.MAX_VALUE);
@@ -583,7 +584,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
             assertThat(bulkShardOperationRequests.get(i), equalTo(Arrays.asList(response.getOperations()).subList(offset, offset + 8)));
         }
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentWrites(), equalTo(32));
     }
 
@@ -610,7 +611,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
             assertThat(operations, equalTo(Arrays.asList(response.getOperations())));
         }
         assertThat(task.isStopped(), equalTo(false));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
         assertThat(status.followerGlobalCheckpoint(), equalTo(-1L));
     }
@@ -638,7 +639,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
             assertThat(operations, equalTo(Arrays.asList(response.getOperations())));
         }
         assertThat(task.isStopped(), equalTo(true));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
         assertThat(status.followerGlobalCheckpoint(), equalTo(-1L));
     }
@@ -660,7 +661,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(bulkShardOperationRequests.size(), equalTo(1));
         assertThat(bulkShardOperationRequests.get(0), equalTo(Arrays.asList(response.getOperations())));
         assertThat(task.isStopped(), equalTo(true));
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentWrites(), equalTo(1));
         assertThat(status.followerGlobalCheckpoint(), equalTo(-1L));
     }
@@ -704,7 +705,7 @@ public class ShardFollowNodeTaskTests extends ESTestCase {
         assertThat(shardChangesRequests.get(0)[0], equalTo(64L));
         assertThat(shardChangesRequests.get(0)[1], equalTo(64L));
 
-        ShardFollowNodeTask.Status status = task.getStatus();
+        ShardFollowNodeTaskStatus status = task.getStatus();
         assertThat(status.numberOfConcurrentReads(), equalTo(1));
         assertThat(status.lastRequestedSeqNo(), equalTo(63L));
         assertThat(status.leaderGlobalCheckpoint(), equalTo(63L));
