+++ x-pack/plugin/ccr/src/test/java/org/elasticsearch/xpack/ccr/action/TransportFollowIndexActionTests.java
@@ -3,6 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+
 package org.elasticsearch.xpack.ccr.action;
 
 import org.elasticsearch.Version;
@@ -15,32 +16,33 @@ import org.elasticsearch.index.mapper.MapperService;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xpack.ccr.CcrSettings;
 import org.elasticsearch.xpack.ccr.ShardChangesIT;
+import org.elasticsearch.xpack.core.ccr.action.FollowIndexAction;
 
 import java.io.IOException;
 
+import static org.elasticsearch.xpack.ccr.action.TransportFollowIndexAction.validate;
 import static org.hamcrest.Matchers.equalTo;
 
-public class FollowIndexActionTests extends ESTestCase {
+public class TransportFollowIndexActionTests extends ESTestCase {
 
     public void testValidation() throws IOException {
         FollowIndexAction.Request request = ShardChangesIT.createFollowRequest("index1", "index2");
         {
             // should fail, because leader index does not exist
-            Exception e = expectThrows(IllegalArgumentException.class, () -> FollowIndexAction.validate(request, null, null, null));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, null, null, null));
             assertThat(e.getMessage(), equalTo("leader index [index1] does not exist"));
         }
         {
             // should fail, because follow index does not exist
             IndexMetaData leaderIMD = createIMD("index1", 5, Settings.EMPTY);
-            Exception e = expectThrows(IllegalArgumentException.class, () -> FollowIndexAction.validate(request, leaderIMD, null, null));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, null, null));
             assertThat(e.getMessage(), equalTo("follow index [index2] does not exist"));
         }
         {
             // should fail because leader index does not have soft deletes enabled
             IndexMetaData leaderIMD = createIMD("index1", 5, Settings.EMPTY);
             IndexMetaData followIMD = createIMD("index2", 5, Settings.EMPTY);
-            Exception e = expectThrows(IllegalArgumentException.class,
-                () -> FollowIndexAction.validate(request, leaderIMD, followIMD, null));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, followIMD, null));
             assertThat(e.getMessage(), equalTo("leader index [index1] does not have soft deletes enabled"));
         }
         {
@@ -48,8 +50,7 @@ public class FollowIndexActionTests extends ESTestCase {
             IndexMetaData leaderIMD = createIMD("index1", 5, Settings.builder()
                 .put(IndexSettings.INDEX_SOFT_DELETES_SETTING.getKey(), "true").build());
             IndexMetaData followIMD = createIMD("index2", 4, Settings.EMPTY);
-            Exception e = expectThrows(IllegalArgumentException.class,
-                () -> FollowIndexAction.validate(request, leaderIMD, followIMD, null));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, followIMD, null));
             assertThat(e.getMessage(),
                 equalTo("leader index primary shards [5] does not match with the number of shards of the follow index [4]"));
         }
@@ -59,8 +60,7 @@ public class FollowIndexActionTests extends ESTestCase {
                 .put(IndexSettings.INDEX_SOFT_DELETES_SETTING.getKey(), "true").build());
             IndexMetaData followIMD = createIMD("index2", State.OPEN, "{}", 5, Settings.builder()
                 .put(IndexSettings.INDEX_SOFT_DELETES_SETTING.getKey(), "true").build());
-            Exception e = expectThrows(IllegalArgumentException.class,
-                () -> FollowIndexAction.validate(request, leaderIMD, followIMD, null));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, followIMD, null));
             assertThat(e.getMessage(), equalTo("leader and follow index must be open"));
         }
         {
@@ -71,8 +71,7 @@ public class FollowIndexActionTests extends ESTestCase {
                 Settings.builder().put(CcrSettings.CCR_FOLLOWING_INDEX_SETTING.getKey(), true).build());
             MapperService mapperService = MapperTestUtils.newMapperService(xContentRegistry(), createTempDir(), Settings.EMPTY, "index2");
             mapperService.updateMapping(null, followIMD);
-            Exception e = expectThrows(IllegalArgumentException.class,
-                () -> FollowIndexAction.validate(request, leaderIMD, followIMD, mapperService));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, followIMD, mapperService));
             assertThat(e.getMessage(), equalTo("mapper [field] of different type, current_type [text], merged_type [keyword]"));
         }
         {
@@ -86,8 +85,7 @@ public class FollowIndexActionTests extends ESTestCase {
                 .put(CcrSettings.CCR_FOLLOWING_INDEX_SETTING.getKey(), true)
                 .put("index.analysis.analyzer.my_analyzer.type", "custom")
                 .put("index.analysis.analyzer.my_analyzer.tokenizer", "standard").build());
-            Exception e = expectThrows(IllegalArgumentException.class,
-                () -> FollowIndexAction.validate(request, leaderIMD, followIMD, null));
+            Exception e = expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, followIMD, null));
             assertThat(e.getMessage(), equalTo("the leader and follower index settings must be identical"));
         }
         {
@@ -100,8 +98,8 @@ public class FollowIndexActionTests extends ESTestCase {
             MapperService mapperService = MapperTestUtils.newMapperService(xContentRegistry(), createTempDir(),
                 followingIndexSettings, "index2");
             mapperService.updateMapping(null, followIMD);
-            IllegalArgumentException error = expectThrows(IllegalArgumentException.class,
-                () -> FollowIndexAction.validate(request, leaderIMD, followIMD, mapperService));
+            IllegalArgumentException error =
+                    expectThrows(IllegalArgumentException.class, () -> validate(request, leaderIMD, followIMD, mapperService));
             assertThat(error.getMessage(), equalTo("the following index [index2] is not ready to follow; " +
                 "the setting [index.xpack.ccr.following_index] must be enabled."));
         }
@@ -113,7 +111,7 @@ public class FollowIndexActionTests extends ESTestCase {
                 .put(CcrSettings.CCR_FOLLOWING_INDEX_SETTING.getKey(), true).build());
             MapperService mapperService = MapperTestUtils.newMapperService(xContentRegistry(), createTempDir(), Settings.EMPTY, "index2");
             mapperService.updateMapping(null, followIMD);
-            FollowIndexAction.validate(request, leaderIMD, followIMD, mapperService);
+            validate(request, leaderIMD, followIMD, mapperService);
         }
         {
             // should succeed, index settings are identical
@@ -129,7 +127,7 @@ public class FollowIndexActionTests extends ESTestCase {
             MapperService mapperService = MapperTestUtils.newMapperService(xContentRegistry(), createTempDir(),
                 followIMD.getSettings(), "index2");
             mapperService.updateMapping(null, followIMD);
-            FollowIndexAction.validate(request, leaderIMD, followIMD, mapperService);
+            validate(request, leaderIMD, followIMD, mapperService);
         }
         {
             // should succeed despite whitelisted settings being different
@@ -147,7 +145,7 @@ public class FollowIndexActionTests extends ESTestCase {
             MapperService mapperService = MapperTestUtils.newMapperService(xContentRegistry(), createTempDir(),
                 followIMD.getSettings(), "index2");
             mapperService.updateMapping(null, followIMD);
-            FollowIndexAction.validate(request, leaderIMD, followIMD, mapperService);
+            validate(request, leaderIMD, followIMD, mapperService);
         }
     }
