+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/XPackClient.java
@@ -13,6 +13,7 @@ import org.elasticsearch.protocol.xpack.XPackInfoRequest;
 import org.elasticsearch.protocol.xpack.XPackInfoResponse;
 import org.elasticsearch.xpack.core.action.XPackInfoAction;
 import org.elasticsearch.xpack.core.action.XPackInfoRequestBuilder;
+import org.elasticsearch.xpack.core.ccr.client.CcrClient;
 import org.elasticsearch.xpack.core.ml.client.MachineLearningClient;
 import org.elasticsearch.xpack.core.monitoring.client.MonitoringClient;
 import org.elasticsearch.xpack.core.security.client.SecurityClient;
@@ -20,6 +21,7 @@ import org.elasticsearch.xpack.core.watcher.client.WatcherClient;
 
 import java.util.Collections;
 import java.util.Map;
+import java.util.Objects;
 
 import static org.elasticsearch.xpack.core.security.authc.support.UsernamePasswordToken.BASIC_AUTH_HEADER;
 import static org.elasticsearch.xpack.core.security.authc.support.UsernamePasswordToken.basicAuthHeaderValue;
@@ -28,6 +30,7 @@ public class XPackClient {
 
     private final Client client;
 
+    private final CcrClient ccrClient;
     private final LicensingClient licensingClient;
     private final MonitoringClient monitoringClient;
     private final SecurityClient securityClient;
@@ -35,7 +38,8 @@ public class XPackClient {
     private final MachineLearningClient machineLearning;
 
     public XPackClient(Client client) {
-        this.client = client;
+        this.client = Objects.requireNonNull(client, "client");
+        this.ccrClient = new CcrClient(client);
         this.licensingClient = new LicensingClient(client);
         this.monitoringClient = new MonitoringClient(client);
         this.securityClient = new SecurityClient(client);
@@ -45,6 +49,10 @@ public class XPackClient {
 
     public Client es() {
         return client;
+    }
+
+    public CcrClient ccr() {
+        return ccrClient;
     }
 
     public LicensingClient licensing() {
